backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
redis>=5.0.0
boto3>=1.34.0
apscheduler>=3.10.0
openpyxl>=3.1.0
weasyprint>=60.0
pydantic-settings>=2.0.0
httpx>=0.25.0
hypothesis>=6.88.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
testcontainers[postgresql]>=3.7.0
moto[ce]>=4.2.0
responses>=0.24.0
